Moves filesystem.h into baldr
so that it is packaged properly as part of `make install`.

Closes https://github.com/valhalla/valhalla/issues/2189
