+++ src/baldr/graphtile.cc
@@ -1,9 +1,9 @@
 #include "baldr/graphtile.h"
 #include "baldr/compression_utils.h"
 #include "baldr/datetime.h"
+#include "baldr/filesystem.h"
 #include "baldr/sign.h"
 #include "baldr/tilehierarchy.h"
-#include "filesystem.h"
 #include "midgard/aabb2.h"
 #include "midgard/pointll.h"
 #include "midgard/tiles.h"
