+++ src/mjolnir/bssbuilder.cc
@@ -19,10 +19,10 @@
 #include <boost/range/algorithm.hpp>
 
 #include "baldr/datetime.h"
+#include "baldr/filesystem.h"
 #include "baldr/graphreader.h"
 #include "baldr/graphtile.h"
 #include "baldr/tilehierarchy.h"
-#include "filesystem.h"
 #include "midgard/distanceapproximator.h"
 #include "midgard/logging.h"
 #include "midgard/sequence.h"
