+++ src/mjolnir/graphfilter.cc
@@ -6,12 +6,12 @@
 #include <unordered_map>
 #include <vector>
 
+#include "baldr/filesystem.h"
 #include "baldr/graphconstants.h"
 #include "baldr/graphid.h"
 #include "baldr/graphreader.h"
 #include "baldr/graphtile.h"
 #include "baldr/tilehierarchy.h"
-#include "filesystem.h"
 #include "midgard/encoded.h"
 #include "midgard/logging.h"
 #include "midgard/pointll.h"
