+++ src/mjolnir/hierarchybuilder.cc
@@ -13,12 +13,12 @@
 #include <utility>
 #include <vector>
 
+#include "baldr/filesystem.h"
 #include "baldr/graphconstants.h"
 #include "baldr/graphid.h"
 #include "baldr/graphreader.h"
 #include "baldr/graphtile.h"
 #include "baldr/tilehierarchy.h"
-#include "filesystem.h"
 #include "midgard/logging.h"
 #include "midgard/pointll.h"
 #include "midgard/sequence.h"
