+++ src/mjolnir/transitbuilder.cc
@@ -16,10 +16,10 @@
 #include <boost/algorithm/string.hpp>
 
 #include "baldr/datetime.h"
+#include "baldr/filesystem.h"
 #include "baldr/graphreader.h"
 #include "baldr/graphtile.h"
 #include "baldr/tilehierarchy.h"
-#include "filesystem.h"
 #include "midgard/distanceapproximator.h"
 #include "midgard/logging.h"
 #include "midgard/util.h"
