+++ src/mjolnir/util.cc
@@ -1,7 +1,7 @@
 #include "mjolnir/util.h"
 
+#include "baldr/filesystem.h"
 #include "baldr/tilehierarchy.h"
-#include "filesystem.h"
 #include "midgard/aabb2.h"
 #include "midgard/logging.h"
 #include "midgard/point2.h"
