+++ src/mjolnir/validatetransit.cc
@@ -14,12 +14,12 @@
 #include <boost/tokenizer.hpp>
 
 #include "baldr/datetime.h"
+#include "baldr/filesystem.h"
 #include "baldr/graphconstants.h"
 #include "baldr/graphid.h"
 #include "baldr/graphreader.h"
 #include "baldr/graphtile.h"
 #include "baldr/tilehierarchy.h"
-#include "filesystem.h"
 #include "midgard/logging.h"
 #include "midgard/sequence.h"
