+++ src/tile_server.cc
@@ -1,6 +1,6 @@
 #include "valhalla/tile_server.h"
 
-#include "valhalla/filesystem.h"
+#include "baldr/filesystem.h"
 
 #include "baldr/compression_utils.h"
