+++ test/astar.cc
@@ -2,12 +2,12 @@
 #include <cstdint>
 #include <fstream>
 
+#include "baldr/filesystem.h"
 #include "baldr/graphid.h"
 #include "baldr/graphreader.h"
 #include "baldr/location.h"
 #include "baldr/rapidjson_utils.h"
 #include "baldr/tilehierarchy.h"
-#include "filesystem.h"
 #include "loki/search.h"
 #include "loki/worker.h"
 #include "midgard/pointll.h"
