+++ test/filesystem.cc
@@ -1,4 +1,4 @@
-#include "filesystem.h"
+#include "baldr/filesystem.h"
 #include "test.h"
 
 #include <algorithm>
